requirements.txt
# Multi-Agent Workflow with Foundry Local - Requirements
# Generated for Python 3.8+ compatibility

# Core AI Framework Dependencies
agent-framework>=0.1.0
openai>=1.0.0,<3.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data Validation and Processing
# Note: Using Pydantic v1 for compatibility with agent-framework
pydantic>=1.10.0,<2.0.0

# HTTP Client Dependencies (for OpenAI/API calls)
httpx>=0.25.0
httpcore>=1.0.0
anyio>=4.0.0
h11>=0.14.0

# Core Python Utilities
typing_extensions>=4.8.0
tqdm>=4.60.0
certifi>=2023.7.22
idna>=3.4
sniffio>=1.3.0
distro>=1.8.0
colorama>=0.4.6

# JSON Processing
jiter>=0.1.0

# Development and Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Installation Notes:
# 1. If you encounter import errors with agent-framework, try:
#    pip install microsoft-agent-framework
#    or
#    pip install azure-agent-framework
#
# 2. For Pydantic compatibility issues, ensure you're using v1.x:
#    pip install "pydantic>=1.10.0,<2.0.0"
#
# 3. If you get dependency conflicts, try installing in this order:
#    pip install pydantic==1.10.24
#    pip install agent-framework
#    pip install -r requirements.txt